@@ -1,7 +1,8 @@
+@file:Suppress("UnstableApiUsage")
+
 plugins {
     id("com.android.application")
     id("kotlin-android")
-    id("kotlin-android-extensions")
 }
 
 repositories {
@@ -9,18 +10,19 @@ repositories {
 }
 
 android {
-    compileSdkVersion(29)
+    compileSdkVersion(31)
     defaultConfig {
         applicationId = "juniojsv.minimum"
-        versionCode = 13
-        versionName = "1.2.2"
+        versionCode = 130
+        versionName = "1.3.0"
         setMinSdkVersion(21)
-        targetSdkVersion(29)
+        targetSdkVersion(31)
         resConfigs("pt-rBr")
         setProperty("archivesBaseName", "$applicationId-v$versionName")
     }
     kotlinOptions {
         jvmTarget = "1.8"
+        useIR = true
     }
     buildTypes {
         getByName("release") {
@@ -29,11 +31,13 @@ android {
             proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
         }
     }
-    buildToolsVersion = "30.0.0"
+    buildFeatures {
+        viewBinding = true
+    }
+    buildToolsVersion = "30.0.3"
 }
 
 dependencies {
     implementation(fileTree(mapOf("dir" to "libs", "include" to listOf("*.jar"))))
     implementation("androidx.preference:preference-ktx:1.1.1")
-    implementation("com.jmedeisis:draglinearlayout:1.1.0")
 }
